fix: fix static analysis
