+++ .gitignore
@@ -7,7 +7,7 @@ nbproject
 .idea
 .project
 .settings
-/vendor
+vendor
 test/coverage-report
 config/autoload/local.php
 .sublime-project
