+++ src/Service/Auth/Exception/RuntimeException.php
@@ -5,6 +5,7 @@
  *
  * @author Rob Caiger <rob@clocal.co.uk>
  */
+
 namespace Dvsa\Olcs\Auth\Service\Auth\Exception;
 
 /**
