+++ test/Service/Auth/PasswordServiceFactoryTest.php
