+++ test/Service/Auth/PasswordServiceTest.php
