--- test/phpunit-coverage.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<phpunit bootstrap="Bootstrap.php" colors="true">
-    <testsuites>
-        <testsuite name="OLCS Auth">
-            <directory>./</directory>
-        </testsuite>
-    </testsuites>
-    <filter>
-        <whitelist processUncoveredFilesFromWhitelist="true">
-            <directory suffix=".php">../src</directory>
-            <exclude>
-                <file>../src/Module.php</file>
-            </exclude>
-        </whitelist>
-    </filter>
-    <logging>
-        <log type="coverage-html" target="./coverage-report" charset="UTF-8"
-             yui="true" highlight="true" lowUpperBound="50"
-             highLowerBound="80"/>
-    </logging>
-</phpunit>
