--- test/phpunit.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" bootstrap="Bootstrap.php" colors="true" xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.3/phpunit.xsd">
-  <coverage processUncoveredFiles="true">
-    <include>
-      <directory suffix=".php">../src</directory>
-    </include>
-  </coverage>
-  <testsuites>
-    <testsuite name="OLCS Auth">
-      <directory>./</directory>
-    </testsuite>
-  </testsuites>
-  <listeners>
-    <listener class="JohnKary\PHPUnit\Listener\SpeedTrapListener">
-      <arguments>
-        <array>
-          <element key="slowThreshold">
-            <integer>250</integer>
-          </element>
-          <element key="reportLength">
-            <integer>5</integer>
-          </element>
-        </array>
-      </arguments>
-    </listener>
-  </listeners>
-</phpunit>
