--- test/src/Service/Auth/Callback/PasswordCallbackTest.php
@@ -1,36 +1,0 @@
-<?php
-
-/**
- * Password Callback Test
- *
- * @author Rob Caiger <rob@clocal.co.uk>
- */
-namespace Dvsa\OlcsTest\Auth\Service\Auth\Callback;
-
-use Dvsa\Olcs\Auth\Service\Auth\Callback\PasswordCallback;
-
-/**
- * Password Callback Test
- *
- * @author Rob Caiger <rob@clocal.co.uk>
- */
-class PasswordCallbackTest extends \PHPunit\Framework\TestCase
-{
-    public function testCallback()
-    {
-        $sut = new PasswordCallback('UserPassword', 'ID1', 'test', false);
-        $result = $sut->toArray();
-        $expected = [
-            'type' => 'PasswordCallback',
-            'output' => [['name' => 'prompt', 'value' => 'UserPassword']],
-            'input' => [
-                [
-                    'name' => 'ID1',
-                    'value' => 'test'
-                ]
-            ]
-        ];
-
-        $this->assertEquals($expected, $result);
-    }
-}
