+++ vendor-bin/phpstan/composer.json
@@ -1,0 +1,5 @@
+{
+  "require-dev": {
+    "phpstan/phpstan": "^1.10"
+  }
+}
