+++ vendor-bin/psalm/composer.json
@@ -1,0 +1,5 @@
+{
+  "require-dev": {
+    "vimeo/psalm": "^5.15"
+  }
+}
